ServiceMP3: Optimize GStreamer start sequence
We start PAUSED and on switch change READY to PAUSED
we are switch to PAYING.

Additionally we do not try to unpause if it is not paused
already.

Finally it will send event start on enigma when state changes
from READY to PAUSED, normally only once and only when stream is valid.

Signed-off-by: Taapat <taapat@gmail.com>